Creates services directory and move sockets file inside
